requirements.txt
aiosqlite==0.20.0
alembic==1.13.1
annotated-types==0.7.0
anyio==3.7.1
appnope==0.1.4
apprise==1.8.0
asgi-lifespan==2.1.0
asn1crypto==1.5.1
asttokens==2.4.1
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
cachetools==5.3.3
certifi==2025.4.26
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
comm==0.2.2
coolname==2.2.0
croniter==2.0.5
cryptography==41.0.7
dateparser==1.2.0
datetime==5.5
debugpy==1.8.1
decorator==5.1.1
dnspython==2.6.1
docker==6.1.3
email_validator==2.1.1
exceptiongroup==1.2.1
executing==2.0.1
filelock==3.15.1
fsspec==2024.6.0
google-auth==2.30.0
google-cloud-secret-manager==2.20.2
google-cloud-storage==2.18.2
graphviz==0.20.3
greenlet==3.0.3
griffe==0.45.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httpx==0.27.0
humanize==4.9.0
hyperframe==6.0.1
idna==3.7
importlib_resources==6.1.3
ipykernel==6.29.4
ipython==8.25.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jinja2-humanize-extension==0.4.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.2
jupyter_core==5.7.2
kubernetes==29.0.0
Mako==1.3.5
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
mdurl==0.1.2
nest-asyncio==1.6.0
numpy==1.26.4
oauthlib==3.2.2
orjson==3.10.5
packaging==24.1
pandas==2.1.4
parso==0.8.4
pathspec==0.12.1
pendulum==2.1.2
pexpect==4.9.0
platformdirs==3.11.0
postmarker==1.0
prefect==2.19.5
prefect-snowflake==0.28.3
prefect-gcp==0.6.2
prompt_toolkit==3.0.47
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==19.0.1
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pydantic==2.7.4
pydantic_core==2.18.4
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==23.3.0
python-dateutil==2.9.0.post0
python-multipart==0.0.9
python-slugify==8.0.4
pytz==2024.1
pytzdata==2020.1
PyYAML==6.0.1
pyzmq==26.0.3
readchar==4.1.0
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==13.7.1
rpds-py==0.18.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
shellingham==1.5.4
simple-salesforce==1.12.6
six==1.16.0
sniffio==1.3.1
snowflake-connector-python==3.15.0
snowflake-sqlalchemy==1.7.1
sortedcontainers==2.4.0
SQLAlchemy==2.0.30
stack-data==0.6.3
text-unidecode==1.3
toml==0.10.2
tomlkit==0.12.5
tornado==6.4.1
traitlets==5.14.3
typer==0.12.3
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
ujson==5.10.0
urllib3==2.4.0
uvicorn==0.28.1
wcwidth==0.2.13
websocket-client==1.8.0
websockets==12.0
